refactor: extract RefreshResult
