+++ extension/src/test/LanguageFunctions.test.ts
@@ -20,6 +20,7 @@ import { RefreshXlfHint, TranslationMode } from "../Enums";
 import * as SettingsLoader from "../Settings/SettingsLoader";
 import { random } from "lodash";
 import { workspace } from "vscode";
+import { RefreshResult } from "../RefreshResult";
 
 const xmlns = "urn:oasis:names:tc:xliff:document:1.2";
 const testResourcesPath = "../../src/test/resources/";
@@ -1212,32 +1213,33 @@ suite("ALObject TransUnit Tests", function () {
 
 suite("Language Functions Tests", function () {
   test("RefreshResult.isChanged()", function () {
-    let refreshResult = new LanguageFunctions.RefreshResult();
+    let refreshResult = new RefreshResult();
     assert.strictEqual(
-      refreshResult.isChanged(),
+      refreshResult.isChanged,
       false,
       "Initialized RefreshResult should not be considered changed"
     );
     refreshResult.numberOfCheckedFiles = 2;
     assert.strictEqual(
-      refreshResult.isChanged(),
+      refreshResult.isChanged,
       false,
       "RefreshResult with numberOfCheckedFiles > 0 should not be considered changed"
     );
     refreshResult.numberOfRemovedNotes = random(1, 1000, false);
     assert.strictEqual(
-      refreshResult.isChanged(),
+      refreshResult.isChanged,
       true,
       "RefreshResult should be considered changed"
     );
-    refreshResult = new LanguageFunctions.RefreshResult();
+    refreshResult = new RefreshResult();
     refreshResult.numberOfReviewsAdded = random(1, 1000, false);
     assert.strictEqual(
-      refreshResult.isChanged(),
+      refreshResult.isChanged,
       true,
       "RefreshResult should be considered changed"
     );
   });
+
   test("LoadMatchXlfIntoMap()", function () {
     /*
      *   - Test with Xlf that has [NAB:* ] tokens
@@ -1868,7 +1870,7 @@ suite("Language Functions Tests", function () {
     </body>
   </file>
 </xliff>`);
-    const refreshResult = new LanguageFunctions.RefreshResult();
+    const refreshResult = new RefreshResult();
     const languageFunctionsSettings = new LanguageFunctionsSettings(
       SettingsLoader.getSettings()
     );
@@ -1923,7 +1925,7 @@ suite("Language Functions Tests", function () {
   });
 
   test("RefreshResult.getReport", function () {
-    const refreshResult = new LanguageFunctions.RefreshResult();
+    const refreshResult = new RefreshResult();
     refreshResult.numberOfAddedTransUnitElements = 1;
     refreshResult.numberOfUpdatedNotes = 1;
     refreshResult.numberOfUpdatedMaxWidths = 1;
@@ -2078,7 +2080,7 @@ function refreshXlfOptionCaptions(
     </body>
   </file>
 </xliff>`);
-  const refreshResult = new LanguageFunctions.RefreshResult();
+  const refreshResult = new RefreshResult();
   const languageFunctionsSettings = new LanguageFunctionsSettings(
     SettingsLoader.getSettings()
   );
